+++ test/testHelper.h
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: Copyright (c) 2022 The ObjCrypto Project Authors 
+// SPDX-FileCopyrightText: Copyright (c) 2022 The ObjCrypto Project Authors
 // SPDX-License-Identifier: BSD-2-Clause
 
 #pragma once
@@ -6,13 +6,12 @@
 #include <iostream>
 
 inline void printHex(const char *name, void *data, size_t size) {
-    uint8_t *ptr = (uint8_t *)data;
+  uint8_t *ptr = (uint8_t *)data;
 
-    std::cout << " " << name << ": ";
-    for (int i = 0; i < size; i++) {
-        if (((i % 4) == 0) && (i > 0))
-            std::cout << "_";
-        std::cout << std::setw(2) << std::setfill('0') << std::hex << (int)(ptr[i]);
-    }
-    std::cout << std::endl;
+  std::cout << " " << name << ": ";
+  for (size_t i = 0; i < size; i++) {
+    if (((i % 4) == 0) && (i > 0)) std::cout << "_";
+    std::cout << std::setw(2) << std::setfill('0') << std::hex << (int)(ptr[i]);
+  }
+  std::cout << std::endl;
 }
